refactor: rename NavbarComponent and associated test file to Navbar
