+++ client/src/App/App.tsx
@@ -1,7 +1,7 @@
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import DashboardContainer from "../containers/FunctionPerformance//DashboardContainer";
 import CloudwatchContainer from "../containers/CloudWatchMetrics/CloudwatchContainer";
-import NavbarComponent from "../components/Navbar/NavbarComponent";
+import NavbarComponent from "../components/Navbar/Navbar";
 import ConfigPageContainer from "../containers/ConfigPage/ConfigPageContainer";
 import ChatContainer from "../containers/FunctionPerformance/Chat/ChatContainer";
 import "../utils/chartSetup";
